Better window selection heuristics. Version bump.
When sending an existing window to the back of the
z-order, we have to choose another window to bring
to the front and activate. Candidate foreground
windows are now chosen by an algorithm similar to
the one used for the system Alt+Tab listing.

We are now more accurate in selecting the window
handles we choose to move/resize -- (what we refer
to as "tangible" windows). Starting from the HWND
that was clicked on, we walk up the window hierarchy
(by parent) until we hit either a popup window or
a window with no parent. The handle we find becomes
the tangible window that we will apply move/resize
operations to.

Remove the ability to send windows to the back of the
z-order by middle-clicking (without holding ALT) on
window title bars. This was old behavior inspired by
the BeOS, but it causes grief when you try to close
tabs in Google Chrome by middle clicking on them. It
is too easy to accidentally activate the send-to-back
feature of Grapple.

Ready for 64-bit builds. I modified the vcproj files
to send 64-bit builds to a separate output directory.
For now though, Grapple defaults to 32-bit only.